:lipstick: reducer 수정 & 분리
