+++ client/src/containers/ChoiceContainer.js
@@ -4,11 +4,11 @@ import { selectAnswerComplete, selectAnswerEvent } from '../actions';
 
 const mapStateToProps = (state, { choice, questionId }) => {
   const number = String.fromCharCode(97 + choice.number);
-  const isSelected = state.reducer.userAnswer[questionId] === number;
+  const isSelected = state.ui.userAnswer[questionId] === number;
   let choiceType = isSelected ? 'selected' : '';
 
-  if (state.reducer.isCheckComplete) {
-    if (state.reducer.checkResult[questionId].correctAnswer === number) {
+  if (state.ui.isCheckComplete) {
+    if (state.ui.checkResult[questionId].correctAnswer === number) {
       choiceType = 'answer';
     } else if (isSelected) {
       choiceType = 'wrong';
@@ -18,8 +18,8 @@ const mapStateToProps = (state, { choice, questionId }) => {
   return {
     choiceType,
     number,
-    questionsLength: state.reducer.questions.length,
-    userAnswer: state.reducer.userAnswer
+    questionsLength: state.content.questions.length,
+    userAnswer: state.ui.userAnswer
   };
 };
