+++ client/src/containers/ChunkContainer.js
@@ -2,7 +2,7 @@ import { connect } from 'react-redux';
 import Chunk from '../components/Chunk/Chunk';
 
 const mapStateToProps = state => {
-  return { chunkMap: state.reducer.chunkMap };
+  return { chunkMap: state.content.chunkMap };
 };
 
 export default connect(mapStateToProps)(Chunk);
